Counter
Added a counter to increase the # displayed in Secrets found every time the secret is found. The mouseover event will off on the secret that is found so that the counter will not count the same secret more than once. Once the counter reached the total number of secrets the counter will stop.